style: separate UIKit and AppKit snapshot tests and outputs
