+++ Tests/PersianJustifyTests/SnapshotTests/UIKitSnapshotTests.swift
@@ -3,7 +3,8 @@ import FontBlaster
 import SnapshotTesting
 @testable import PersianJustify
 
-final class SnapshotTests: XCTestCase {
+#if canImport(UIKit)
+final class UIKitSnapshotTests: XCTestCase {
     override class func setUp() {
         FontBlaster.blast(bundle: .test)
     }
@@ -12,10 +13,7 @@ final class SnapshotTests: XCTestCase {
         let fonts = FontBlaster.loadedFonts
         XCTAssertFalse(fonts.isEmpty)
     }
-}
 
-#if canImport(UIKit)
-extension SnapshotTests {
     func testSampleViewController() throws {
         let sut = SampleViewController()
 
@@ -32,26 +30,6 @@ extension SnapshotTests {
             sut.numberOfLines = 0
             sut.textColor = .black
             sut.attributedText = justifiedText
-            assertSnapshot(
-                of: sut,
-                as: .image(size: CGSize(width: width, height: 780)),
-                named: "Font:\(fontName)"
-            )
-        }
-    }
-}
-#endif
-
-#if canImport(AppKit)
-extension SnapshotTests {
-    func testLongMultilineTextOnNSTextField() throws {
-        let width: CGFloat = 360
-        for fontName in FontBlaster.loadedFonts {
-            let font = try XCTUnwrap(Font(name: fontName, size: 17))
-            let justifiedText = longDemoText.toPJString(fittingWidth: width, font: font)
-            let sut = NSTextField(labelWithAttributedString: justifiedText)
-            sut.textColor = .black
-
             assertSnapshot(
                 of: sut,
                 as: .image(size: CGSize(width: width, height: 780)),
